[C] (+) rename waitlist->steamder
