+++ packages/client/src/hooks/useGetSteamder.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { getSteamder } from '../services/api/waitlists/get';
+import { getSteamder } from '../services/api/steamders/get';
 
 export const useGetSteamder = (steamderId: string) => {
     return useMutation({ mutationFn: () => getSteamder(steamderId), mutationKey: ['steamder', steamderId], retry: false });
