+++ packages/client/src/hooks/useJoinSteamder.ts
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import { joinSteamder } from "../services/api/waitlists/join";
+import { joinSteamder } from "../services/api/steamders/join";
 
 export const useJoinSteamder = (steamderId: string) => {
     return useMutation({ mutationFn: () => joinSteamder(steamderId), mutationKey: ['steamder', steamderId], retry: false });
